ov065: promote daObjCtMecha03_c into a single C++ TU
Gathers the eight one-function shards that held this Tick Tock Clock
pendulum into src/actors/daObjCtMecha03_c.cpp, the translation unit the
cartridge shipped.

    python tools/tubuild.py verify ov065/daObjCtMecha03_c
    8/8 MATCH, objisolate clean, reloc-destinations clean -> TEXT-VERIFIED

    python tools/tubuild.py linkcheck ov065/daObjCtMecha03_c
    SCRATCH-LINK-VERIFIED. 0x02119f3c..0x0211a45c IDENTICAL, module
    fidelity 106/106 exact, ROM sha256 d1506e90... identical to the
    pristine-tree baseline control. dsd check symbols --fail still fails
    on the same 9 pre-existing errors the baseline has: 0 new.

Destructor placement. The cartridge puts D1 at 0x02119f3c ahead of D0 at
0x02119f88, which is exactly what mwccarm 2004/b56 emits from an inline
in-class virtual destructor. include/daObjCtMecha03_c.h therefore moves
its declaration to `virtual ~daObjCtMecha03_c() {}` as the first member.
That is the only header change; no field was renamed or respelled, so no
field-address CSE lever moved underneath a byte-verified body.

compiler_only_output is 12 rows, not the queue's estimated ~13. The
estimate was copied from daObjCtMecha05_c; the closed form for this
inheritance depth is 2 x (4 ancestors + self) + 1 vtable + 1 Vector3 D1
= 12, and the measured object agrees exactly. Every canonical_module and
canonical_address was read out of the ROM's own symbols.txt rather than
from the sibling manifest. Two rows are NOT in this TU's module:
_ZTI10dBgActor_c and _ZTS10dBgActor_c have vague linkage and the linker
kept their single copy in ov002 (0x021089ec, 0x02108a04).

_ZN7Vector3D1Ev is the one deadstrip-duplicate: the genuine dBgCh_Gnd
local in func_ov065_02119fe8 makes mwcc emit Vector3's empty inline
destructor, whose canonical home is arm9 0x020072c0.

func_ov065_02119fe8 and func_ov065_0211a114 keep their address-derived
names because ov065's symbols.txt spells them that way; they stay
extern "C" free functions rather than being renamed into members.

All three `tubuild create` CONFLICT notes were parameter-name or
u8-versus-unsigned-char spelling differences over the same types; each
is recorded as RESOLVED in the manifest with the resolution appended.

Bookkeeping: 8 delinks.txt entries spliced into 1 (CRLF preserved),
manifest status -> promoted, tiers_ratchet --update moved the three
member rows onto the promoted path, attribution overrides re-keyed onto
the surviving path, and notes/bgobject-provenance.md no longer names the
deleted shards.
