--- src/_ZN16daObjCtMecha03_c6RenderEv.cpp
@@ -1,11 +1,0 @@
-//cpp
-// @symbol _ZN16daObjCtMecha03_c6RenderEv
-/* recovered: named members + shared header, real C++ method */
-#include "daObjCtMecha03_c.h"
-struct Base { virtual void v0(); virtual void v1(); virtual void v2(); virtual void v3(); virtual void v4(); virtual void m(int); };
-struct Derived { char pad[0xd4]; Base base; };
-
-int daObjCtMecha03_c::Render()
-{
- Base *b = &((Derived *)this)->base; b->m(0); return 1;
-}
